+++ src/utils/segments-middleware.js
@@ -1,7 +1,5 @@
 import _ from "lodash";
 
-const fieldPath = "ship.private_settings.synchronized_segments";
-
 /**
  * @param  {Object}   req
  * @param  {Object}   res
@@ -15,7 +13,7 @@ export default function segmentsMiddlewareFactory() {
       return next();
     }
 
-    const { cache, message } = req.hull;
+    const { cache, message, connectorConfig } = req.hull;
     const bust = (message
       && (message.Subject === "users_segment:update" || message.Subject === "users_segment:delete"));
 
@@ -28,6 +26,8 @@ export default function segmentsMiddlewareFactory() {
       return cache.wrap("segments", () => req.hull.client.get("/segments"));
     }).then((segments) => {
       req.hull.segments = _.map(segments, (s) => {
+        const fieldName = connectorConfig.segmentFilterSetting;
+        const fieldPath = `ship.private_settings.${fieldName}`;
         if (_.has(req.hull, fieldPath)) {
           s.filtered = _.includes(_.get(req.hull, fieldPath, []), s.id);
         }